move meinberlinIdeaCollection to core
src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/IdeaCollection.ts
src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/Process/Process.ts
