+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/AddProposalButton.html
