+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/DetailColumn.html
@@ -11,11 +11,11 @@
         </div>
     </a>
 
-    <adh-meinberlin-idea-collection-detail
+    <adh-idea-collection-detail
         data-ng-switch-when="body"
         data-ng-if="processUrl"
         data-path="{{processUrl}}"
         data-is-buergerhaushalt="isBuergerhaushalt"
         data-is-kiezkasse="isKiezkasse">
-    </adh-meinberlin-idea-collection-detail>
+    </adh-idea-collection-detail>
 </div>
