+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/EditColumn.html
@@ -2,11 +2,11 @@
     <div data-ng-switch-when="menu">
     </div>
 
-    <adh-meinberlin-idea-collection-edit
+    <adh-idea-collection-edit
         data-ng-switch-when="body"
         data-ng-if="processUrl"
         data-path="{{processUrl}}"
         data-is-buergerhaushalt="isBuergerhaushalt"
         data-is-kiezkasse="isKiezkasse">
-    </adh-meinberlin-idea-collection-edit>
+    </adh-idea-collection-edit>
 </div>
