+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/Process/Detail.html
@@ -21,11 +21,11 @@
                 data-path="{{path}}"
                 data-sort="sort"
                 data-sorts="sorts">
-                <adh-meinberlin-proposal-list-item
+                <adh-idea-collection-proposal-list-item
                     data-path="{{element}}"
                     data-is-buergerhaushalt="custom.isBuergerhaushalt"
                     data-is-kiezkasse="custom.isKiezkasse">
-                </adh-meinberlin-proposal-list-item>
+                </adh-idea-collection-proposal-list-item>
             </adh-listing>
 
             <div
@@ -50,11 +50,11 @@
                     data-polygon="polygon"
                     data-sort="sort"
                     data-sorts="sorts">
-                    <adh-meinberlin-proposal-map-list-item
+                    <adh-idea-collection-proposal-map-list-item
                         data-path="{{item}}"
                         data-is-buergerhaushalt="custom.isBuergerhaushalt"
                         data-is-kiezkasse="custom.isKiezkasse">
-                    </adh-meinberlin-proposal-map-list-item>
+                    </adh-idea-collection-proposal-map-list-item>
                 </adh-map-listing>
             </adh-wait>
         </div>
