+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/Process/Module.ts
@@ -1,0 +1,18 @@
+import * as AdhHttpModule from "../../Http/Module";
+import * as AdhPermissionsModule from "../../Permissions/Module";
+
+import * as Process from "./Process";
+
+
+export var moduleName = "adhIdeaCollectionProcess";
+
+export var register = (angular) => {
+    angular
+        .module(moduleName, [
+            AdhHttpModule.moduleName,
+            AdhPermissionsModule.moduleName
+        ])
+        .directive("adhIdeaCollectionDetail", ["adhConfig", "adhHttp", "adhPermissions", "$q", Process.detailDirective])
+        .directive("adhIdeaCollectionEdit", [
+            "adhConfig", "adhHttp", "adhShowError", "adhSubmitIfValid", "moment", Process.editDirective]);
+};
