+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/ProcessHeaderSlot.html
@@ -1,0 +1,1 @@
+<adh-idea-collection-add-proposal-button></adh-idea-collection-add-proposal-button>
