+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/ProposalCreateColumn.html
@@ -9,10 +9,10 @@
         </div>
     </div>
 
-    <adh-meinberlin-proposal-create
+    <adh-idea-collection-proposal-create
         data-pool-path="{{processUrl}}"
         data-is-buergerhaushalt="isBuergerhaushalt"
         data-is-kiezkasse="isKiezkasse"
         data-ng-switch-when="body" >
-    </adh-meinberlin-proposal-create>
+    </adh-idea-collection-proposal-create>
 </div>
