+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/ProposalDetailColumn.html
@@ -28,12 +28,12 @@
                 data-hide="true"
                 data-print="true">
             </adh-resource-actions>
-            <adh-meinberlin-proposal-detail
+            <adh-idea-collection-proposal-detail
                 data-ng-if="proposalUrl"
                 data-is-buergerhaushalt="isBuergerhaushalt"
                 data-is-kiezkasse="isKiezkasse"
                 data-path="{{proposalUrl}}">
-            </adh-meinberlin-proposal-detail>
+            </adh-idea-collection-proposal-detail>
         </adh-recompile-on-change>
     </div>
 </div>
