+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/IdeaCollection/ProposalEditColumn.html
@@ -17,11 +17,11 @@
     <adh-recompile-on-change
             data-ng-switch-when="body"
             data-value="{{proposalUrl}}">
-        <adh-meinberlin-proposal-edit
+        <adh-idea-collection-proposal-edit
             data-ng-if="proposalUrl"
             data-is-buergerhaushalt="isBuergerhaushalt"
             data-is-kiezkasse="isKiezkasse"
             data-path="{{proposalUrl}}">
-        </adh-meinberlin-proposal-edit>
+        </adh-idea-collection-proposal-edit>
     </adh-recompile-on-change>
 </div>
