+++ src/meinberlin/meinberlin/static/js/Adhocracy.ts
@@ -36,6 +36,7 @@ import * as AdhDoneModule from "./Packages/Done/Module";
 import * as AdhEmbedModule from "./Packages/Embed/Module";
 import * as AdhEventManagerModule from "./Packages/EventManager/Module";
 import * as AdhHttpModule from "./Packages/Http/Module";
+import * as AdhIdeaCollectionModule from "./Packages/IdeaCollection/Module";
 import * as AdhImageModule from "./Packages/Image/Module";
 import * as AdhInjectModule from "./Packages/Inject/Module";
 import * as AdhListingModule from "./Packages/Listing/Module";
@@ -102,6 +103,7 @@ export var init = (config : AdhConfig.IService, metaApi) => {
         AdhCrossWindowMessagingModule.moduleName,
         AdhDebateWorkbenchModule.moduleName,
         AdhEmbedModule.moduleName,
+        AdhIdeaCollectionModule.moduleName,
         AdhMeinberlinModule.moduleName,
         AdhResourceAreaModule.moduleName,
         AdhTrackingModule.moduleName,
@@ -191,6 +193,7 @@ export var init = (config : AdhConfig.IService, metaApi) => {
     AdhEmbedModule.register(angular);
     AdhEventManagerModule.register(angular);
     AdhHttpModule.register(angular, config);
+    AdhIdeaCollectionModule.register(angular);
     AdhImageModule.register(angular);
     AdhInjectModule.register(angular);
     AdhListingModule.register(angular);
