+++ src/meinberlin/meinberlin/static/js/Packages/Meinberlin/Alexanderplatz/Workbench/ProposalCreateColumn.html
@@ -16,6 +16,6 @@
             data-resource-path="{{processUrl}}"
             data-cancel="true">
         </adh-resource-actions>
-        <adh-meinberlin-proposal-create data-pool-path="{{processUrl}}"></adh-meinberlin-proposal-create>
+        <adh-idea-collection-proposal-create data-pool-path="{{processUrl}}"></adh-idea-collection-proposal-create>
     </div>
 </div>
