+++ src/meinberlin/meinberlin/static/js/Packages/Meinberlin/Alexanderplatz/Workbench/ProposalDetailColumn.html
@@ -25,9 +25,9 @@
             data-resource-path="{{proposalUrl | adhParentPath}}"
             data-edit="true">
         </adh-resource-actions>
-        <adh-meinberlin-proposal-detail
+        <adh-idea-collection-proposal-detail
             data-ng-if="proposalUrl"
             data-path="{{proposalUrl}}">
-        </adh-meinberlin-proposal-detail>
+        </adh-idea-collection-proposal-detail>
     </adh-recompile-on-change>
 </div>
