+++ src/meinberlin/meinberlin/static/js/Packages/Meinberlin/Alexanderplatz/Workbench/ProposalEditColumn.html
@@ -20,9 +20,9 @@
             data-resource-path="{{proposalUrl | adhParentPath}}"
             data-cancel="true">
         </adh-resource-actions>
-        <adh-meinberlin-proposal-edit
+        <adh-idea-collection-proposal-edit
             data-ng-if="proposalUrl"
             data-path="{{proposalUrl}}">
-        </adh-meinberlin-proposal-edit>
+        </adh-idea-collection-proposal-edit>
     </adh-recompile-on-change>
 </div>
