--- src/meinberlin/meinberlin/static/js/Packages/Meinberlin/IdeaCollection/ProcessHeaderSlot.html
@@ -1,1 +1,0 @@
-<adh-meinberlin-idea-collection-add-proposal-button></adh-meinberlin-idea-collection-add-proposal-button>
