+++ src/meinberlin/meinberlin/static/js/Packages/Meinberlin/Kiezkasse/Module.ts
@@ -2,13 +2,13 @@ import * as AdhEmbedModule from "../../Embed/Module";
 import * as AdhProcessModule from "../../Process/Module";
 import * as AdhResourceAreaModule from "../../ResourceArea/Module";
 
-import * as AdhMeinberlinIdeaCollectionModule from "../IdeaCollection/Module";
+import * as AdhIdeaCollectionModule from "../../IdeaCollection/Module";
 
 import * as AdhEmbed from "../../Embed/Embed";
 import * as AdhProcess from "../../Process/Process";
 import * as AdhResourceArea from "../../ResourceArea/ResourceArea";
 
-import * as AdhMeinberlinIdeaCollection from "../IdeaCollection/IdeaCollection";
+import * as AdhIdeaCollection from "../../IdeaCollection/IdeaCollection";
 
 import RIKiezkasseProcess from "../../../Resources_/adhocracy_meinberlin/resources/kiezkassen/IProcess";
 
@@ -21,7 +21,7 @@ export var register = (angular) => {
     angular
         .module(moduleName, [
             AdhEmbedModule.moduleName,
-            AdhMeinberlinIdeaCollectionModule.moduleName,
+            AdhIdeaCollectionModule.moduleName,
             AdhProcessModule.moduleName,
             AdhResourceAreaModule.moduleName
         ])
@@ -29,15 +29,15 @@ export var register = (angular) => {
             adhEmbedProvider.registerContext("kiezkasse", ["kiezkassen"]);
         }])
         .config(["adhResourceAreaProvider", "adhConfig", (adhResourceAreaProvider : AdhResourceArea.Provider, adhConfig) => {
-            var registerRoutes = AdhMeinberlinIdeaCollection.registerRoutesFactory(processType);
+            var registerRoutes = AdhIdeaCollection.registerRoutesFactory(processType);
             registerRoutes(processType)(adhResourceAreaProvider);
             registerRoutes(processType, "kiezkasse")(adhResourceAreaProvider);
 
-            var processHeaderSlot = adhConfig.pkg_path + AdhMeinberlinIdeaCollection.pkgLocation + "/ProcessHeaderSlot.html";
+            var processHeaderSlot = adhConfig.pkg_path + AdhIdeaCollection.pkgLocation + "/ProcessHeaderSlot.html";
             adhResourceAreaProvider.processHeaderSlots[processType] = processHeaderSlot;
         }])
         .config(["adhProcessProvider", (adhProcessProvider : AdhProcess.Provider) => {
             adhProcessProvider.templates[processType] =
-                "<adh-meinberlin-idea-collection-workbench data-is-kiezkasse=\"true\"></adh-meinberlin-idea-collection-workbench>";
+                "<adh-idea-collection-workbench data-is-kiezkasse=\"true\"></adh-idea-collection-workbench>";
         }]);
 };
