+++ src/meinberlin/meinberlin/static/js/Packages/Meinberlin/Module.ts
@@ -2,9 +2,7 @@ import * as AdhMeinberlinAlexanderplatzModule from "./Alexanderplatz/Module";
 import * as AdhMeinberlinBplanModule from "./Bplan/Module";
 import * as AdhMeinberlinBuergerhaushaltModule from "./Buergerhaushalt/Module";
 import * as AdhMeinberlinDeModule from "./MeinberlinDe/Module";
-import * as AdhMeinberlinIdeaCollectionModule from "./IdeaCollection/Module";
 import * as AdhMeinberlinKiezkasseModule from "./Kiezkasse/Module";
-import * as AdhMeinberlinProposalModule from "./Proposal/Module";
 import * as AdhMeinberlinStadtforumModule from "./Stadtforum/Module";
 
 
@@ -15,9 +13,7 @@ export var register = (angular) => {
     AdhMeinberlinBplanModule.register(angular);
     AdhMeinberlinBuergerhaushaltModule.register(angular);
     AdhMeinberlinDeModule.register(angular);
-    AdhMeinberlinIdeaCollectionModule.register(angular);
     AdhMeinberlinKiezkasseModule.register(angular);
-    AdhMeinberlinProposalModule.register(angular);
     AdhMeinberlinStadtforumModule.register(angular);
 
     angular
@@ -26,9 +22,7 @@ export var register = (angular) => {
             AdhMeinberlinBplanModule.moduleName,
             AdhMeinberlinBuergerhaushaltModule.moduleName,
             AdhMeinberlinDeModule.moduleName,
-            AdhMeinberlinIdeaCollectionModule.moduleName,
             AdhMeinberlinKiezkasseModule.moduleName,
-            AdhMeinberlinProposalModule.moduleName,
             AdhMeinberlinStadtforumModule.moduleName
         ]);
 };
